+++ application/backend/src/business/services/releases/helpers/release-data-egress-helper.ts
