+++ application/backend/tests/service-tests/audits/audit-event-user.test.ts
@@ -2,7 +2,7 @@ import { AuthenticatedUser } from "../../../src/business/authenticated-user";
 import { AuditEventService } from "../../../src/business/services/audit-event-service";
 import * as edgedb from "edgedb";
 import { registerTypes } from "../../test-dependency-injection.common";
-import { beforeEachCommon } from "../user.common";
+import { beforeEachCommon } from "../commons/user.common";
 import { addSeconds } from "date-fns";
 
 let existingUser: AuthenticatedUser;
