+++ application/backend/tests/service-tests/manifests/manifest.master.test.ts
@@ -1,5 +1,5 @@
 import { Client } from "edgedb";
-import { beforeEachCommon } from "../releases/releases.common";
+import { beforeEachCommon } from "../commons/releases.common";
 import { registerTypes } from "../../test-dependency-injection.common";
 import { ManifestService } from "../../../src/business/services/manifests/manifest-service";
 import { AuthenticatedUser } from "../../../src/business/authenticated-user";
