+++ application/backend/tests/service-tests/rems.test.ts
@@ -1,5 +1,5 @@
 import { AuthenticatedUser } from "../../src/business/authenticated-user";
-import { beforeEachCommon } from "./releases/releases.common";
+import { beforeEachCommon } from "./commons/releases.common";
 import { registerTypes } from "../test-dependency-injection.common";
 import { RemsService } from "../../src/business/services/dacs/rems-service";
 import {
